refactor out plugins
